hopefully fixes canonicalization bug that rarely collapses triples & more checks
Before this change, the reduction to an RDF BNode canonicalization
problem would actually not transform fixed vars into BNodes, but they
ended up as rdflib.Variables in the rdflib.Graph. While this seems to
be possible, it's not a well tested area of rdflib. This is why we now
transform fixed vars into URIRefs with a specific prefix and then
transform them back after solution

Additionally, we introduce further not too expensive checks that will
check that canonicalization did not collapse triples, or modify certain
features (degrees, counts, etc). If it does, a warning is issued and we
fall back to returning the non canonicalized gp.