multi_query error temporarily closes down worker
multi_query raises errors in case of severe problems only. In those
cases cases continued querying causes more problems. With this commit
an error that is raised out of multi_query will cause the worker to
temporarily be closed. Further calls to multi_query will be logged as
warning and raise a MultiQueryException. This allows the scoop main
process to early terminate and handle the exception.

Waiting times are now configurable and in general increased to 30s as
they are more direct now. This for example allows the endpoint to
restart in case of buuuuugs.